chore: remove obsolete authentication pages and add new layout for protected routes
