--- src/app/(auth)/register/page.tsx
@@ -1,11 +1,0 @@
-import { RegisterForm } from "@/components/register-form";
-
-export default function RegisterPage() {
-  return (
-    <div className="flex min-h-svh flex-col items-center justify-center bg-muted p-6 md:p-10">
-      <div className="w-full max-w-sm md:max-w-3xl">
-        <RegisterForm />
-      </div>
-    </div>
-  );
-}
