--- src/app/(auth)/verify-email/page.tsx
@@ -1,109 +1,0 @@
-"use client";
-
-import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { useSignUp } from "@clerk/nextjs";
-import { useState } from "react";
-import { useRouter } from "next/navigation";
-
-export default function VerifyEmailPage() {
-  const { signUp, isLoaded, setActive } = useSignUp();
-  const [code, setCode] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState("");
-  const router = useRouter();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    
-    if (!isLoaded) return;
-    
-    try {
-      setIsLoading(true);
-      setError("");
-      
-      const result = await signUp.attemptEmailAddressVerification({
-        code,
-      });
-      
-      if (result.status === "complete") {
-        await setActive({ session: result.createdSessionId });
-        router.push("/dashboard");
-      } else {
-        console.error("Verification failed", result);
-        setError("Verification failed. Please try again.");
-      }
-    } catch (err: any) {
-      console.error("Error during verification:", err);
-      setError(err.errors?.[0]?.message || "Failed to verify email. Please try again.");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleResendCode = async () => {
-    if (!isLoaded) return;
-    
-    try {
-      await signUp.prepareEmailAddressVerification({ strategy: "email_code" });
-      alert("Verification code has been resent to your email");
-    } catch (err) {
-      console.error("Error resending code:", err);
-      setError("Failed to resend verification code");
-    }
-  };
-
-  return (
-    <div className="flex min-h-svh flex-col items-center justify-center bg-muted p-6 md:p-10">
-      <div className="w-full max-w-sm md:max-w-md">
-        <Card>
-          <CardContent className="p-6">
-            <div className="flex flex-col gap-6">
-              <div className="flex flex-col items-center text-center">
-                <h1 className="text-2xl font-bold">Verify your email</h1>
-                <p className="text-balance text-muted-foreground mt-2">
-                  We've sent a verification code to your email address. Please enter it below.
-                </p>
-              </div>
-              
-              {error && (
-                <div className="bg-destructive/15 text-destructive text-sm p-3 rounded-md">
-                  {error}
-                </div>
-              )}
-              
-              <form onSubmit={handleSubmit} className="flex flex-col gap-4">
-                <div className="grid gap-2">
-                  <Label htmlFor="code">Verification Code</Label>
-                  <Input
-                    id="code"
-                    value={code}
-                    onChange={(e) => setCode(e.target.value)}
-                    placeholder="Enter code"
-                    required
-                  />
-                </div>
-                
-                <Button type="submit" disabled={isLoading}>
-                  {isLoading ? "Verifying..." : "Verify Email"}
-                </Button>
-              </form>
-              
-              <div className="text-center">
-                <button 
-                  onClick={handleResendCode}
-                  className="text-sm text-primary hover:underline"
-                  type="button"
-                >
-                  Didn't receive a code? Resend
-                </button>
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-      </div>
-    </div>
-  );
-}
