+++ src/app/(protected)/(sidebar)/layout.tsx
@@ -1,4 +1,5 @@
 import { AppSidebar } from "@/components/app-sidebar";
+import { ModeToggle } from "@/components/theme-switcher";
 import {
     Breadcrumb,
     BreadcrumbItem,
@@ -24,13 +25,13 @@ export default function SidebarLayout({
             <AppSidebar />
             <SidebarInset>
                 <header className="flex h-16 shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[[data-collapsible=icon]]/sidebar-wrapper:h-12">
-                    <div className="flex items-center gap-2 px-4">
+                    <div className="flex items-center justify-between w-full px-4">
                         <SidebarTrigger className="-ml-1" />
-                        <Separator
+                        {/*<Separator
                             orientation="vertical"
                             className="mr-2 h-4"
                         />
-                        <Breadcrumb>
+                         <Breadcrumb>
                             <BreadcrumbList>
                                 <BreadcrumbItem className="hidden md:block">
                                     <BreadcrumbLink href="#">
@@ -44,7 +45,8 @@ export default function SidebarLayout({
                                     </BreadcrumbPage>
                                 </BreadcrumbItem>
                             </BreadcrumbList>
-                        </Breadcrumb>
+                        </Breadcrumb> */}
+                        <ModeToggle />
                     </div>
                 </header>
                 {children}
